+++ __init__.py
@@ -1,20 +1,9 @@
 """netports"""
 
 from netports.interface4 import Interface4
-from netports.ports import (
-    all_itcp,
-    all_ivlan,
-    all_stcp,
-    all_svlan,
-    iports,
-    itcp,
-    ivlan,
-    ivlan_hpe,
-    ranges,
-    sports,
-    stcp,
-    svlan,
-    svlan_hpe,
-)
+from netports.ip import IP_NAMES, IP_PORTS, iip, iip_all, iip_nip, sip, sip_all
+from netports.ports import inumbers, ranges, snumbers
 from netports.range import Range
 from netports.ranges import Ranges
+from netports.tcp import itcp_all, stcp_all, stcp, itcp
+from netports.vlan import ivlan_all, svlan_all, ivlan, ivlan_hpe, svlan, svlan_hpe