+++ netports/helpers.py
@@ -1,11 +1,17 @@
 """Helper functions"""
 from typing import Any
 
-from netports.types_ import LStr
+from netports.static import SPLITTER
+from netports.types_ import LAny, LStr
+
+
+def join(items: LAny) -> str:
+    """Join items by "," """
+    return SPLITTER.join([str(i) for i in items])
 
 
 def list_of_str(items: Any) -> LStr:
-    """Convert any types of items to list of string"""
+    """Convert any types of items to *List[str]*"""
     if not isinstance(items, (bytes, dict, float, int, list, set, str, tuple)):
         raise TypeError(f"{items=} {list} expected")
     if isinstance(items, (bytes, float, int, str)):
@@ -19,3 +25,12 @@ def list_of_str(items: Any) -> LStr:
         else:
             items_.append(str(item))
     return items_
+
+
+def split(items: Any) -> LStr:
+    """Split items *str*, *List[str]* by "," and <space>"""
+    results: LStr = []
+    for items_s in list_of_str(items):
+        for item_s in items_s.split():
+            results.extend(item_s.split(SPLITTER))
+    return results