+++ netports/ranges.py
@@ -11,17 +11,20 @@ from netports.types_ import LStr, LInt, OInt, SInt, IInt
 class Ranges:
     """Ranges"""
 
-    __slots__ = ("_line", "_ports", "splitter", "range_splitter", "_valid_chars")
+    __slots__ = ("_line", "_numbers", "splitter", "range_splitter", "_valid_chars", "_strict")
 
-    def __init__(self, line: str, splitter: str = SPLITTER, range_splitter: str = RANGE_SPLITTER):
+    def __init__(self, line: str, **kwargs):
         """Ranges.
         :param line: Range in <str> format.
         :param splitter: Separator character between numbers. By default ",".
         :param range_splitter: Separator between min and max digits in range. By default "-".
+        :param strict: True - Raise ValueError, if in line is invalid item.
+                       False - Make Range without invalid items. By default True.
         """
-        self.splitter = splitter
-        self.range_splitter = range_splitter
-        self._ports: LInt = []
+        self.splitter = kwargs.get("splitter") or SPLITTER
+        self.range_splitter = kwargs.get("range_splitter") or RANGE_SPLITTER
+        self._strict = self._init_strict(**kwargs)
+        self._numbers: LInt = []
         self.line = line
 
     def __str__(self):
@@ -56,6 +59,18 @@ class Ranges:
 
     # ============================= init =============================
 
+    @staticmethod
+    def _init_strict(**kwargs) -> bool:
+        """Init strict"""
+        strict = kwargs.get("strict")
+        if strict is None:
+            return True
+        if isinstance(strict, bool):
+            return strict
+        raise TypeError(f"{strict=} {bool} expected")
+
+    # =========================== property ===========================
+
     @property
     def line(self) -> str:
         """Ranges in <str> format"""
@@ -68,17 +83,17 @@ class Ranges:
         line_: str = self._valid_line(line)
         line_ = line_.replace(self.range_splitter, RANGE_SPLITTER)
         line_ = line_.replace(self.splitter, SPLITTER)
-        items = line_.split(SPLITTER)
+        items: LStr = line_.split(SPLITTER)
         items = [s for s in items if s]
-        ranges: LRange = sorted([Range(s) for s in set(items)])
-        ports: SInt = {i for o in ranges for i in o.range}
-        self._ports = sorted(ports)
-        self._line = self._sports(ports)
+        ranges = self._ranges(items)
+        numbers: SInt = {i for o in ranges for i in o.range}
+        self._numbers = sorted(numbers)
+        self._line = self._snumbers(numbers)
 
     @property
-    def ports(self) -> LInt:
+    def numbers(self) -> LInt:
         """Ranges in ist of <int> format"""
-        return self._ports
+        return self._numbers
 
     # =========================== helpers ============================
 
@@ -93,12 +108,14 @@ class Ranges:
             items_ = [s for s in items_ if s]
             for item in items_:
                 if not item.isdigit():
-                    raise ValueError(f"invalid {item=} in {line=}, expected {splitter=}")
+                    if not self._strict:
+                        continue
+                    raise ValueError(f"invalid {item=} in {line=}")
             line_ = splitter.join(items_)
             lines.append(line_)
         return range_splitter.join(lines)
 
-    def _sports(self, items: IInt) -> str:
+    def _snumbers(self, items: IInt) -> str:
         """Convert list of <int> to <str>.
         :param items: [1, 3, 4, 5]
         :return: "1,3-5"
@@ -127,18 +144,15 @@ class Ranges:
                 ranges.append(item_)
         return self.splitter.join(ranges)
 
-
-# ============================= helpers ==============================
-
-def check_tcp_range(ports: LInt) -> bool:
-    """True if all ports are in the valid TCP/UDP range 1...65535, else raise ERROR."""
-    if invalid_vlans := [i for i in ports if i < 1 or i > 65535]:
-        raise ValueError(f"{invalid_vlans=}, expected in range 1...65535")
-    return True
-
-
-def check_vlans_range(vlans: LInt) -> bool:
-    """True if all vlans are in the valid range 1...4094, else raise ERROR."""
-    if invalid_vlans := [i for i in vlans if i < 1 or i > 4094]:
-        raise ValueError(f"{invalid_vlans=}, expected in range 1...4094")
-    return True
+    def _ranges(self, items: LStr) -> LRange:
+        """Convert items to list of Range.
+        If self._strict==True, raise ValueError, else skip invalid item."""
+        ranges: LRange = []
+        for item in items:
+            try:
+                ranges.append(Range(item))
+            except ValueError as ex:
+                if self._strict:
+                    raise type(ex)(*ex.args)
+                continue
+        return ranges