+++ netports/types_.py
@@ -1,6 +1,7 @@
 """Typing"""
 from typing import (
     Any,
+    Dict,
     Iterable,
     List,
     Optional,
@@ -8,6 +9,8 @@ from typing import (
     Tuple,
 )
 
+DAny = Dict[str, Any]
+DiAny = Dict[int, Any]
 IInt = Iterable[int]
 IStr = Iterable[str]
 LAny = List[Any]
@@ -15,5 +18,6 @@ LInt = List[int]
 LStr = List[str]
 OInt = Optional[int]
 SInt = Set[int]
+SStr = Set[str]
 
 TStr5 = Tuple[str, str, str, str, str]